requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-celery-beat==2.0.0
django-cors-headers==3.5.0
django-timezone-field==4.0
djangorestframework==3.12.1
flower==0.9.5
humanize==3.0.1
kombu==4.6.11
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
rabbitmq==0.2.0
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
vine==1.3.0
wcwidth==0.2.5
wincertstore==0.2
